# Improve time performance