--- a/lib/Data/Time/Clock/Scale.hs
+++ b/lib/Data/Time/Clock/Scale.hs
@@ -9,7 +9,7 @@ module Data.Time.Clock.Scale
     UniversalTime(..),
 
     -- * Absolute intervals
-    DiffTime,
+    DiffTime(..),
     secondsToDiffTime,
     picosecondsToDiffTime,
     diffTimeToPicoseconds,