--- a/lib/Data/Time/Format/Parse.hs
+++ b/lib/Data/Time/Format/Parse.hs
@@ -1,5 +1,4 @@
 {-# OPTIONS -fno-warn-orphans #-}
-#include "HsConfigure.h"
 
 -- #hide
 module Data.Time.Format.Parse