Updated, This patch:

+ Change `POSIXTime` representation to a record with two `Int64` field instead of previous `Integer`, which improve many IO operations.

+ Add benchmarks to easily benchmark against system `time`, e.g. the old version.

+ Cleaned up some CPPs and style to reduce further burden.