@winterland1989 2016-11-28
Note there's thunk leak here so we didn't really create any `Integer` in `getCurrentTime`, if we mark `UTCTime` 's field to strict, running time reach to 80ns,  which is still much better than previous `getPOSIXTime` based one.