@hvr 2016-11-28
@winterland1989 just out of curiosity (if you have time/interest to try it), how does the performance compare to e.g https://hackage.haskell.org/package/uhttpc-0.1.1.0/docs/Network-HTTP-MicroClient.html#v:getPOSIXTimeUSecs ?