@winterland1989 2016-11-28
Let me give it a try ; ) 

```
benchmarking time/uhttp
time                 135.3 ns   (134.0 ns .. 136.7 ns)
                     0.999 R²   (0.999 R² .. 0.999 R²)
mean                 136.0 ns   (134.9 ns .. 137.3 ns)
std dev              3.880 ns   (3.264 ns .. 4.816 ns)
variance introduced by outliers: 43% (moderately inflated)
```

It seems IEEE `Double` is no faster on my machine.