@winterland1989 2016-11-29
After trying to optimize `getZonedTime`, now i strongly proposed we either add 

```
data POSIXTimeRaw = POSIXTimeRaw 
    { utSeconds :: {-# UNPACK #-} !CTime
    , utNanoSeconds :: {-# UNPACK #-} !Int64
    }
```

or change `POSIXTime` directly. The reason is that if we don't have this type as bridge and we don't want convert to/from `Integer` during `getZonedTime`, then we were force to include `HsTimeConfig.h` and duplicate code in `Data.Time.Clock.POSIX`. Which is not optimal.