@hvr 2016-11-29
wouldn't changing `POSIXTime` directly result in conversion functions becoming partial?