@winterland1989 2016-11-29
It is indeed a valid concern, new data type is necessary. what about reuse code in `clock` package? The `TimeSpec` is exactly what we need here.