@winterland1989 2016-11-29
I have manage to add `POSIXTimeRaw` and used it to improve many `IO` operations, before:

```
benchmarking time/getCurrentTime
time                 1.052 μs   (1.041 μs .. 1.065 μs)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 1.053 μs   (1.044 μs .. 1.064 μs)
std dev              31.67 ns   (25.04 ns .. 42.80 ns)
variance introduced by outliers: 41% (moderately inflated)

benchmarking time/getPOSIXTime
time                 248.9 ns   (247.0 ns .. 251.0 ns)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 251.6 ns   (248.9 ns .. 254.4 ns)
std dev              8.611 ns   (7.456 ns .. 10.57 ns)
variance introduced by outliers: 51% (severely inflated)

benchmarking time/getZonedTime
time                 4.659 μs   (4.620 μs .. 4.694 μs)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 4.654 μs   (4.609 μs .. 4.708 μs)
std dev              164.2 ns   (132.0 ns .. 223.0 ns)
variance introduced by outliers: 45% (moderately inflated)
```

now:

```
benchmarking time/getCurrentTime
time                 85.67 ns   (84.61 ns .. 86.77 ns)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 85.39 ns   (84.61 ns .. 86.34 ns)
std dev              2.816 ns   (2.216 ns .. 3.556 ns)
variance introduced by outliers: 51% (severely inflated)

benchmarking time/getPOSIXTime
time                 129.1 ns   (128.3 ns .. 130.2 ns)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 128.5 ns   (127.4 ns .. 129.6 ns)
std dev              3.602 ns   (3.043 ns .. 4.343 ns)
variance introduced by outliers: 42% (moderately inflated)

benchmarking time/getZonedTime
time                 869.1 ns   (861.1 ns .. 877.0 ns)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 866.5 ns   (859.3 ns .. 874.5 ns)
std dev              25.94 ns   (21.97 ns .. 31.13 ns)
variance introduced by outliers: 41% (moderately inflated)
```