@winterland1989 2016-11-29
There're lots of CPP, code and document clean up shoud be done, i'll push more commits after reviewing ; )