@erikd 2016-11-29
I'm always a bit suspicious of benchmarks where the variance is more than about 20%. Are you running these benchmarks on a laptop? If so, have you tried running them on a desktop or server machine?