@winterland1989 2016-11-30
Here's the result from our server(CentOS 7.1):

before my patch:

```
benchmarking time/getCurrentTime
time                 1.831 μs   (1.824 μs .. 1.838 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.832 μs   (1.827 μs .. 1.848 μs)
std dev              28.91 ns   (12.49 ns .. 55.04 ns)
variance introduced by outliers: 15% (moderately inflated)

benchmarking time/getPOSIXTime
time                 329.0 ns   (328.0 ns .. 330.2 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 328.9 ns   (328.3 ns .. 330.6 ns)
std dev              3.299 ns   (1.561 ns .. 6.396 ns)

benchmarking time/getZonedTime
time                 13.94 μs   (13.89 μs .. 13.99 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 13.94 μs   (13.91 μs .. 14.03 μs)
std dev              166.7 ns   (80.97 ns .. 301.1 ns)
```

After:

```
benchmarking time/getCurrentTime
time                 81.97 ns   (81.96 ns .. 81.98 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 81.97 ns   (81.96 ns .. 81.99 ns)
std dev              44.64 ps   (30.14 ps .. 67.80 ps)

benchmarking time/getPOSIXTime
time                 154.9 ns   (154.8 ns .. 154.9 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 154.9 ns   (154.8 ns .. 155.0 ns)
std dev              215.2 ps   (79.73 ps .. 457.8 ps)

benchmarking time/getZonedTime
time                 3.837 μs   (3.828 μs .. 3.847 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 3.837 μs   (3.833 μs .. 3.843 μs)
std dev              17.58 ns   (14.03 ns .. 22.65 ns)
```