@AshleyYakeley 2016-12-01
This is close. However, it contains a number of different changes combined into one PR. I might push some of those changes separately just to make the others easier to examine.