@winterland1989 2016-12-01
Btw, I have a commit add benchmark based on `PackageImports`, i add a new cabal file in `benchmark` which import both current modules and system `time`, it makes benchmark much easier, do you want me to push it now or wait for you finish current patches?