@winterland1989 2016-12-01
An confusing thing just happened, at after mark all kinds of strict field, `formatTime` became slower than original one:

```
benchmarking format/old formatTime
time                 11.80 μs   (11.67 μs .. 11.91 μs)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 11.82 μs   (11.71 μs .. 11.96 μs)
std dev              404.8 ns   (328.8 ns .. 511.2 ns)
variance introduced by outliers: 41% (moderately inflated)

benchmarking format/now formatTime
time                 15.07 μs   (14.91 μs .. 15.28 μs)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 15.15 μs   (14.99 μs .. 15.30 μs)
std dev              532.1 ns   (446.9 ns .. 666.5 ns)
```

+ If i remove strict and unpack pragmas from `LocalTime/TimeOfDay/UTCTime`, `formatTime` become as fast as original one again.

+ If i add strict bang to `UTCTime`, `formatTime` will be a litter faster than original one. 

Is there something tricky happened when ghc try to optimize `formatTime`? Please shed some light.