@winterland1989 2016-12-01
FYI, my benchmark suggest to remove the strcit bang in `LocalTime/TimeOfDay/UTCTime` and keep `TimeZone/POSIXTime` 's strcit and unpack mark. It seems strict bang in these high level type prevent some rules/optimization to fire.