@winterland1989 2016-12-02
Pushed new bench code, you can toggle between #ff9488f and #fd3053b to see how strictness affect `formatTime`.

BTW, i'm working on a rewrite of `formatTime/FormatTime`, early result is very promising, i'll make new PR after this one get merged.