@winterland1989 2016-12-04
I decided that the formatting work will be continued in a separated package. Is there anything i can do on current PR to push it forward?