@AshleyYakeley 2016-12-04
There are a number of additional changes that make it awkward, as well as formatting changes and so on. It's not clear to me that I can remove `HsConfigure.h`, for example. As it is, I'm going to have to cherry-pick it.